requirements.txt
langchain==0.3.26
flask==3.1.1
sentence-transformers==2.7.0
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone==0.2.8
langchain-openai==0.3.24
langchain-community==0.3.26
openai>=1.0.0
pinecone-client>=4.0.0
- e.


langchain==0.3.27
langchain_core==0.3.74
langchain_community==0.3.27
langchain-ollama==0.2.1
langserve==0.3.1
fastapi==0.115.6
starlette==0.41.3
pydantic==2.8.2
sse-starlette==1.8.2
httpx==0.27.2
uvicorn==0.30.6
python-dotenv
streamlit
bs4
pypdf
chromadb
faiss-cpu
LanceDB
wikipedia
arxiv
langchainhub 